refactor(admin): group operator chat console under its own module
The chat view where operators talk to agents, watch tool calls and approve commands had its component, state composable and styles spread across view/, composables/ and styles/. Fold them into admin/chat so the messaging UI evolves as one cohesive unit.
