+++ webui/src/admin/chat/chat.scss
